로그인 반영 완료

+++ src/main/java/com/app/pickcourse/controller/FeedsController.java
@@ -7,6 +7,7 @@ import com.app.pickcourse.domain.vo.ReportVO;
 import com.app.pickcourse.service.FeedsService;
 import com.app.pickcourse.util.PaginationOnePage;
 import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -24,7 +25,9 @@ import java.util.Map;
 @RequiredArgsConstructor
 @Slf4j
 public class FeedsController {
+    private final HttpSession session;
     private final FeedsService feedsService;
+    private final HttpServletRequest request;
 
     // 댓글 목록 조회 :: 25.03.16 조승찬
     @GetMapping("/reply-list/{feedId}")
@@ -32,13 +35,13 @@ public class FeedsController {
                                @PathVariable Long feedId, PaginationOnePage pagination, Model model) {
 
         if (member == null) {
+            session.setAttribute("redirectAfterLogin", request.getRequestURI());
             return "redirect:/login/login";
         }
 
         Long memberId = member.getId();
 
-
-
+        // memberId는 댓글 "삭제", "신고" 를 구분하기 위해서 사용
         List<ReplyListDTO> replys = feedsService.getReplyList(memberId, feedId, pagination);
         log.info("pagination  "+pagination.toString());
         model.addAttribute("replys", replys);  // 댓글 목록
@@ -52,9 +55,6 @@ public class FeedsController {
     @ResponseBody
     public ResponseEntity<Map<String, Object>> getReplyListApi(@SessionAttribute(name = "member", required = false) MemberDTO member,
                                                                @PathVariable Long feedId, PaginationOnePage pagination) {
-        if (member == null) {
-            return ResponseEntity.badRequest().build();
-        }
 
         List<ReplyListDTO> replys = feedsService.getReplyList(member.getId(), feedId, pagination);
         Map<String, Object> response = new HashMap<>();
@@ -67,7 +67,8 @@ public class FeedsController {
     // 댓글 삭제 레스트컨트롤러 방식 => 컨트롤러 방식으로 redirect :: 25.03.16 조승찬
     @DeleteMapping("/reply-list/{id}")
     @ResponseBody
-    public Map<String, Object> deleteReplyList(@PathVariable Long id,
+    public Map<String, Object> deleteReplyList(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                               @PathVariable Long id,
                                                @RequestParam Long feedId) {
 
         feedsService.deleteReplyList(id);
@@ -79,17 +80,19 @@ public class FeedsController {
     // 댓글 신고 25.03.16 조승찬
     @PostMapping("/reply-list/report")
     @ResponseBody
-    public void postReportReplyList(@RequestBody ReportVO reportVO) {
+    public void postReportReplyList(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                    @RequestBody ReportVO reportVO) {
 
-        feedsService.postReportReplyList(reportVO, 1l);  // 로그인수정
+        feedsService.postReportReplyList(reportVO, member.getId());
     }
 
     // 댓글 등록  25.03.17 조승찬
     @PostMapping("/reply-list")
     @ResponseBody
-    public Map<String, Object> postReplyList(@RequestBody ReplyVO replyVO) {
-        log.info("Map<String, Object> postReplyList에 들어옴 "+replyVO.toString());
-        feedsService.postReplyList(replyVO, 1l);   // 로그인수정
+    public Map<String, Object> postReplyList(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                             @RequestBody ReplyVO replyVO) {
+
+        feedsService.postReplyList(replyVO, member.getId());
 
         Map<String, Object> response = new HashMap<>();
         // 댓글 등록 후 첫 페이지로 이동 :: 등록 확인을 위해
@@ -100,9 +103,15 @@ public class FeedsController {
 
     // 나의 댓글 목록 조회 :: 25.03.17 조승찬
     @GetMapping("/my/reply-list")
-    public String getMyReplyList(PaginationOnePage pagination, Model model) {
+    public String getMyReplyList(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                 PaginationOnePage pagination, Model model) {
 
-        List<ReplyListDTO> replys = feedsService.getMyReplyList(1l, pagination);  //로그인수정
+        if (member == null) {
+            session.setAttribute("redirectAfterLogin", request.getRequestURI());
+            return "redirect:/login/login";
+        }
+
+        List<ReplyListDTO> replys = feedsService.getMyReplyList(member.getId(), pagination);
         model.addAttribute("replys", replys);  // 댓글 목록
         model.addAttribute("replyAction", new ReplyActionDTO()); // 입력될 댓글을 받아올 객체
         model.addAttribute("pagination", pagination);
@@ -113,9 +122,10 @@ public class FeedsController {
     // 나의 댓글 목록 추가 조회 레스트컨트롤러 방식:: 25.03.17 조승찬
     @GetMapping("/my/reply-list/api")
     @ResponseBody
-    public ResponseEntity<Map<String, Object>> getMyReplyListApi(PaginationOnePage pagination) {
+    public ResponseEntity<Map<String, Object>> getMyReplyListApi(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                                                 PaginationOnePage pagination) {
 
-        List<ReplyListDTO> replys = feedsService.getMyReplyList(1l, pagination); // 로그인수정
+        List<ReplyListDTO> replys = feedsService.getMyReplyList(member.getId(), pagination);
         Map<String, Object> response = new HashMap<>();
         response.put("replys", replys);
         response.put("pagination", pagination);
@@ -136,7 +146,14 @@ public class FeedsController {
 
     // 피드 작성용 화면 랜딩 25.03.18 조승찬
     @GetMapping("/feed-write")
-    public String getFeedWrite(Model model) {
+    public String getFeedWrite(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                               Model model) {
+
+        if (member == null) {
+            session.setAttribute("redirectAfterLogin", request.getRequestURI());
+            return "redirect:/login/login";
+        }
+
         FeedDTO feedDTO = new FeedDTO();
         model.addAttribute("feedDTO", feedDTO);
         return "/feeds/feed-write";
@@ -144,9 +161,15 @@ public class FeedsController {
 
     // 피드 작성 25.03.18 조승찬
     @PostMapping("/feed-write")
-    public String postFeedWrite(FeedDTO feedDTO) {
+    public String postFeedWrite(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                FeedDTO feedDTO) {
 
-        feedsService.postFeedWrite(1l, feedDTO); //로그인수정
+        if (member == null) {
+            session.setAttribute("redirectAfterLogin", request.getRequestURI());
+            return "redirect:/login/login";
+        }
+
+        feedsService.postFeedWrite(member.getId(), feedDTO);
 
         // 조회 타입 정하기
         String listType = feedDTO.getFeedType().equals("TOGETHER") ? "TOGETHER" : "ALL";
@@ -156,7 +179,13 @@ public class FeedsController {
 
     // 피드 수정용 조회 25.03.19 조승찬
     @GetMapping("/feed-modify")
-    public String getFeedModify(@RequestParam("id") Long id, @RequestParam("feedType") String feedType, Model model) {
+    public String getFeedModify(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                @RequestParam("id") Long id, @RequestParam("feedType") String feedType, Model model) {
+
+        if (member == null) {
+            session.setAttribute("redirectAfterLogin", request.getRequestURI());
+            return "redirect:/login/login";
+        }
 
         FeedDTO feedDTO = feedsService.getFeedModify(id, feedType);
         feedDTO.setId(id);
@@ -168,9 +197,15 @@ public class FeedsController {
 
     // 피드 수정 25.03.19 조승찬
     @PostMapping("/feed-modify")
-    public String postFeedModify(FeedDTO feedDTO) {
+    public String postFeedModify(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                 FeedDTO feedDTO) {
 
-        feedsService.postFeedModify(feedDTO); //로그인수정
+        if (member == null) {
+            session.setAttribute("redirectAfterLogin", request.getRequestURI());
+            return "redirect:/login/login";
+        }
+
+        feedsService.postFeedModify(feedDTO);
 
         String listType = feedDTO.getFeedType().equals("TOGETHER") ? "TOGETHER" : "ALL";
         return "redirect:/feeds/my/feed-list?listType=" + listType;
@@ -178,9 +213,12 @@ public class FeedsController {
 
     // 피드 삭제 25.03.19 조승찬
     @PostMapping("/feed-delete")
-    public String deleteFeedModify(@RequestParam("id") Long id, @RequestParam("feedType") String feedType) {
-
-        log.info("deleteFeedModify 들어옴 ::  "+id+" "+feedType);
+    public String deleteFeedModify(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                   @RequestParam("id") Long id, @RequestParam("feedType") String feedType) {
+        if (member == null) {
+            session.setAttribute("redirectAfterLogin", request.getRequestURI());
+            return "redirect:/login/login";
+        }
 
         feedsService.deleteFeedModify(id, feedType);
 
@@ -190,7 +228,14 @@ public class FeedsController {
 
     // 리얼 후기 작성 25.03.18 조승찬
     @GetMapping("/real-write")
-    public String ReviewWrite(@RequestParam("planId") Long planId, Model model) {
+    public String ReviewWrite(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                              @RequestParam("planId") Long planId, Model model) {
+
+        if (member == null) {
+            session.setAttribute("redirectAfterLogin", request.getRequestURI());
+            return "redirect:/login/login";
+        }
+
         RealDTO realDTO = new RealDTO();
         realDTO.setPlanId(planId);
         model.addAttribute("realDTO", realDTO);
@@ -199,17 +244,26 @@ public class FeedsController {
 
     // 리얼 후기 작성 25.03.18 조승찬
     @PostMapping("/real-write")
-    public String postRealWrite(RealDTO realDTO) {
-        log.info(realDTO.toString());
+    public String postRealWrite(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                RealDTO realDTO) {
+        if (member == null) {
+            session.setAttribute("redirectAfterLogin", request.getRequestURI());
+            return "redirect:/login/login";
+        }
 
-        feedsService.postRealWrite(1l, realDTO); //로그인수정
+        feedsService.postRealWrite(member.getId(), realDTO);
 
         return "redirect:/feeds/feed-list?listType=REAL";
     }
 
     // 리얼 후기 수정용 조회 25.03.19 조승찬
     @GetMapping("/real-modify")
-    public String getRealModify(@RequestParam("id") Long id, Model model) {
+    public String getRealModify(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                @RequestParam("id") Long id, Model model) {
+        if (member == null) {
+            session.setAttribute("redirectAfterLogin", request.getRequestURI());
+            return "redirect:/login/login";
+        }
 
         RealDTO realDTO = feedsService.getRealModify(id);
         realDTO.setId(id);
@@ -220,7 +274,12 @@ public class FeedsController {
 
     // 리얼 후기 수정 25.03.19 조승찬
     @PostMapping("/real-modify")
-    public String postRealModify(RealDTO realDTO) {
+    public String postRealModify(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                 RealDTO realDTO) {
+        if (member == null) {
+            session.setAttribute("redirectAfterLogin", request.getRequestURI());
+            return "redirect:/login/login";
+        }
 
         feedsService.postRealModify(realDTO);
 
@@ -229,7 +288,12 @@ public class FeedsController {
 
     // 리얼 후기 삭제 25.03.19 조승찬
     @PostMapping("/real-delete")
-    public String deleteFeedModify(@RequestParam("id") Long id) {
+    public String deleteFeedModify(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                   @RequestParam("id") Long id) {
+        if (member == null) {
+            session.setAttribute("redirectAfterLogin", request.getRequestURI());
+            return "redirect:/login/login";
+        }
 
         feedsService.deleteRealModify(id);
 
@@ -239,9 +303,10 @@ public class FeedsController {
     // 피드 신고 25.03.20 조승찬
     @PostMapping("/feed-list/report")
     @ResponseBody
-    public void postReportFeedList(@RequestBody ReportVO reportVO) {
+    public void postReportFeedList(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                   @RequestBody ReportVO reportVO) {
 
-        feedsService.postReportFeedList(reportVO, 1l);  // 로그인수정
+        feedsService.postReportFeedList(reportVO, member.getId());
     }
 
 
@@ -257,8 +322,14 @@ public class FeedsController {
 
     // 나의 피드 리스트 25.03.20 조승찬
     @GetMapping("/my/feed-list")
-    public String getFeedModifyList(@RequestParam("listType") String listType, Model model) {
-        List<FeedListDTO> feedListDTO = feedsService.getMyFeedList(1l, listType); //  로그인수정
+    public String getFeedModifyList(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                    @RequestParam("listType") String listType, Model model) {
+        if (member == null) {
+            session.setAttribute("redirectAfterLogin", request.getRequestURI());
+            return "redirect:/login/login";
+        }
+
+        List<FeedListDTO> feedListDTO = feedsService.getMyFeedList(member.getId(), listType);
         model.addAttribute("feedListDTO", feedListDTO);
         model.addAttribute("listType", listType);
         return "/feeds/my-feed-list";
@@ -266,9 +337,14 @@ public class FeedsController {
 
     // 나의 여행 목록 25.03.21 조승찬
     @GetMapping("/tour-list")
-    public String getTourList(PaginationOnePage pagination, Model model) {
+    public String getTourList(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                              PaginationOnePage pagination, Model model) {
+        if (member == null) {
+            session.setAttribute("redirectAfterLogin", request.getRequestURI());
+            return "redirect:/login/login";
+        }
 
-        List<TourListDTO> tourListDTO = feedsService.getTourList(1l, pagination); // 로그인 수정
+        List<TourListDTO> tourListDTO = feedsService.getTourList(member.getId(), pagination);
         model.addAttribute("tourListDTO", tourListDTO);
         model.addAttribute("pagination", pagination);
 
@@ -282,9 +358,10 @@ public class FeedsController {
     // 나의 여행 목록 추가 조회 레스트컨트롤러 방식:: 25.03.21 조승찬
     @GetMapping("/tour-list/api")
     @ResponseBody
-    public ResponseEntity<Map<String, Object>> getTourListApi(PaginationOnePage pagination) {
+    public ResponseEntity<Map<String, Object>> getTourListApi(@SessionAttribute(name = "member", required = false) MemberDTO member,
+                                                              PaginationOnePage pagination) {
 
-        List<TourListDTO> tours = feedsService.getTourList(1l, pagination);  //로그인수정
+        List<TourListDTO> tours = feedsService.getTourList(member.getId(), pagination);
         log.info("레스트 "+pagination.toString());
         tours.forEach(System.out::println);
